+++ src/app/production/[id]/page.tsx
@@ -75,6 +75,7 @@ export default function ProductionConfiguration({ params }: PageProps) {
   >();
   const [createStream, loadingCreateStream] = useCreateStream();
   const [deleteStream, loadingDeleteStream] = useDeleteStream();
+
   //PRODUCTION
   const putProduction = usePutProduction();
   const getPresets = useGetPresets();
