+++ src/components/modal/configureOutputModal/ConfigureOutputModal.tsx
@@ -1,4 +1,4 @@
-import { TMultiviewLayout, Preset } from '../../../interfaces/preset';
+import { Preset } from '../../../interfaces/preset';
 import { Modal } from '../Modal';
 import Decision from './Decision';
 import PipelineOutputConfig, { PipelineTypes } from './PipelineOutputConfig';
@@ -6,20 +6,12 @@ import { useState } from 'react';
 import { PipelineOutput, PipelineSettings } from '../../../interfaces/pipeline';
 import { usePipelines } from '../../../hooks/pipelines';
 import cloneDeep from 'lodash.clonedeep';
-import toast from 'react-hot-toast';
-import { MultiviewSettings } from '../../../interfaces/multiview';
-import MultiviewSettingsConfig from './MultiviewSettings';
-import MultiviewLayoutSettings from './MultiviewLayoutSettings/MultiviewLayoutSettings';
-import { IconPlus, IconTrash } from '@tabler/icons-react';
-import { Production } from '../../../interfaces/production';
-import { usePutMultiviewLayout } from '../../../hooks/multiviewLayout';
 
 type ConfigureOutputModalProps = {
   open: boolean;
   preset: Preset;
   onClose: () => void;
   updatePreset: (preset: Preset) => void;
-  production: Production | undefined;
 };
 
 export interface OutputStream {
@@ -41,8 +33,7 @@ export function ConfigureOutputModal({
   open,
   preset,
   onClose,
-  updatePreset,
-  production
+  updatePreset
 }: ConfigureOutputModalProps) {
   const [pipelines, setPipelines] = useState<PipelineSettings[]>(
     preset.pipelines || []
@@ -52,39 +43,6 @@ export function ConfigureOutputModal({
     useState<number>(DEFAULT_PORT_MUMBER);
 
   const [pipes] = usePipelines();
-  const [multiviews, setMultiviews] = useState<MultiviewSettings[]>([]);
-  const [portDuplicateIndexes, setPortDuplicateIndexes] = useState<number[]>(
-    []
-  );
-  const [layoutModalOpen, setLayoutModalOpen] = useState(false);
-  const [selectedMultiviewLayout, setSelectedMultiviewLayout] = useState<
-    TMultiviewLayout | undefined
-  >();
-  const [newMultiviewLayout, setNewMultiviewLayout] =
-    useState<TMultiviewLayout | null>(null);
-  const addNewLayout = usePutMultiviewLayout();
-  const t = useTranslate();
-
-  useEffect(() => {
-    if (preset.pipelines[0].multiviews) {
-      if (!Array.isArray(preset.pipelines[0].multiviews)) {
-        setMultiviews([preset.pipelines[0].multiviews]);
-      } else {
-        setMultiviews(preset.pipelines[0].multiviews);
-      }
-    }
-  }, [preset.pipelines]);
-
-  useEffect(() => {
-    setOutputStreams(defaultState(preset.pipelines));
-  }, [preset]);
-
-  const clearInputs = () => {
-    setLayoutModalOpen(false);
-    setMultiviews(preset.pipelines[0].multiviews || []);
-    setOutputStreams(defaultState(preset.pipelines));
-    onClose();
-  };
 
   const onSave = () => {
     const locations = pipelines
@@ -109,126 +67,18 @@ export function ConfigureOutputModal({
     onClose();
   };
 
-  const onUpdateLayoutPreset = () => {
-    const noLayoutName = newMultiviewLayout?.name === '';
-    const defaultLayout = newMultiviewLayout?.name.includes('Default');
-    if (!newMultiviewLayout || noLayoutName || defaultLayout) {
-      toast.error(t('preset.no_updated_layout'));
-      return;
-    }
-    addNewLayout(newMultiviewLayout);
-    setLayoutModalOpen(false);
-  };
-
-  const closeLayoutModal = () => {
-    setLayoutModalOpen(false);
-  };
-
-  const streamsToProgramOutputs = (
-    pipelineIndex: number,
-    outputStreams?: OutputStream[]
+  const updatePipelineOutputFunc = (
+    pipeline: PipelineSettings,
+    outputs: PipelineOutput[]
   ) => {
-    if (!outputStreams) return [];
-    return outputStreams.map((stream) => ({
-      ...preset.pipelines[pipelineIndex].program_output[0],
-      port: stream.port,
-      [stream.srtMode === 'listener' ? 'local_ip' : 'remote_ip']: stream.ip,
-      srt_mode: stream.srtMode,
-      video_bit_depth: stream.videoBit,
-      video_format: stream.videoFormat,
-      video_kilobit_rate: stream.videoKiloBit,
-      srt_passphrase: stream.srtPassphrase
-    })) satisfies ProgramOutput[];
-  };
-
-  const addStream = (stream: OutputStream) => {
-    const streams = outputstreams.filter(
-      (o) => o.pipelineIndex === stream.pipelineIndex
+    setCurrentError('');
+    const pipelineIndex = pipelines.findIndex(
+      (p) => p.pipeline_name === pipeline.pipeline_name
     );
-    if (streams.length > 4) return;
-    setOutputStreams([
-      ...outputstreams,
-      {
-        ...stream,
-        name: `${t('preset.stream_name')} ${streams.length + 1}`,
-        port: streams[streams.length - 1].port + 1
-      }
-    ]);
-  };
-
-  const updateStream = (updatedStream: OutputStream) => {
-    setOutputStreams(
-      [
-        ...outputstreams.filter((o) => o.id !== updatedStream.id),
-        updatedStream
-      ].sort((a, b) => a.name.localeCompare(b.name))
-    );
-  };
-
-  const updateStreams = (updatedStreams: OutputStream[]) => {
-    const streams = outputstreams.filter(
-      (o) => !updatedStreams.some((u) => u.id === o.id)
-    );
-    setOutputStreams(
-      [...streams, ...updatedStreams].sort((a, b) =>
-        a.name.localeCompare(b.name)
-      )
-    );
-  };
-
-  const setNames = (outputstreams: OutputStream[], index: number) => {
-    const streamsForPipe = outputstreams.filter(
-      (o) => o.pipelineIndex === index
-    );
-    const rest = outputstreams.filter((o) => o.pipelineIndex !== index);
-    return [
-      ...streamsForPipe.map((s, i) => ({ ...s, name: `Stream ${i + 1}` })),
-      ...rest
-    ];
-  };
-
-  const deleteStream = (id: string, index: number) => {
-    setOutputStreams(
-      setNames(
-        outputstreams.filter((o) => o.id !== id),
-        index
-      )
-    );
-  };
-
-  const findDuplicateValues = (mvs: MultiviewSettings[]) => {
-    const ports = mvs.map(
-      (item: MultiviewSettings) =>
-        item.output.local_ip + ':' + item.output.local_port.toString()
-    );
-    const duplicateIndices: number[] = [];
-    const seenPorts = new Set();
-
-    ports.forEach((port, index) => {
-      if (seenPorts.has(port)) {
-        duplicateIndices.push(index);
-        // Also include the first occurrence if it's not already included
-        const firstIndex = ports.indexOf(port);
-        if (!duplicateIndices.includes(firstIndex)) {
-          duplicateIndices.push(firstIndex);
-        }
-      } else {
-        seenPorts.add(port);
-      }
-    });
-
-    return duplicateIndices;
-  };
-
-  const runDuplicateCheck = (mvs: MultiviewSettings[]) => {
-    const hasDuplicates = findDuplicateValues(mvs);
-
-    if (hasDuplicates.length > 0) {
-      setPortDuplicateIndexes(hasDuplicates);
-    }
-
-    if (hasDuplicates.length === 0) {
-      setPortDuplicateIndexes([]);
+    if (pipelineIndex >= 0) {
+      const newPipelines = cloneDeep(pipelines);
+      newPipelines.splice(pipelineIndex, 1, { ...pipeline, outputs: outputs });
+      setPipelines(newPipelines);
     }
   };
 
@@ -237,35 +87,11 @@ export function ConfigureOutputModal({
     return currentPortNumber;
   };
 
-  const handleUpdateMultiview = (
-    multiview: MultiviewSettings,
-    index: number
-  ) => {
-    const updatedMultiviews = multiviews.map((item, i) =>
-      i === index ? { ...item, ...multiview } : item
-    );
-
-    runDuplicateCheck(multiviews);
-
-    setMultiviews(updatedMultiviews);
-  };
-
-  const addNewMultiview = (newMultiview: MultiviewSettings) => {
-    setMultiviews((prevMultiviews) =>
-      prevMultiviews ? [...prevMultiviews, newMultiview] : [newMultiview]
-    );
-  };
-
-  const removeNewMultiview = (index: number) => {
-    const newMultiviews = multiviews.filter((_, i) => i !== index);
-    setMultiviews(newMultiviews);
-  };
-
   return (
     <Modal
       open={open}
       outsideClick={() => {
-        clearInputs();
+        onClose();
       }}
     >
       <div className="overflow-auto max-h-[90vh]">
@@ -288,79 +114,9 @@ export function ConfigureOutputModal({
               />
             );
           })}
-          {multiviews &&
-            multiviews.length > 0 &&
-            multiviews.map((singleItem, index) => {
-              return (
-                <div className="flex" key={index}>
-                  <div className="min-h-full border-l border-separate opacity-10 my-12"></div>
-                  <div className="flex flex-col">
-                    <MultiviewSettingsConfig
-                      productionId={production?._id}
-                      openConfigModal={() => setLayoutModalOpen(true)}
-                      newMultiviewLayout={newMultiviewLayout}
-                      lastItem={multiviews.length === index + 1}
-                      multiview={singleItem}
-                      handleUpdateMultiview={(input) =>
-                        handleUpdateMultiview(input, index)
-                      }
-                      portDuplicateError={
-                        portDuplicateIndexes.length > 0
-                          ? portDuplicateIndexes.includes(index)
-                          : false
-                      }
-                      setSelectedMultiviewLayout={(layout) =>
-                        setSelectedMultiviewLayout(layout)
-                      }
-                      selectedMultiviewLayout={selectedMultiviewLayout}
-                    />
-                    <div
-                      className={`w-full flex ${
-                        multiviews.length > 1
-                          ? 'justify-between'
-                          : 'justify-end'
-                      }`}
-                    >
-                      {multiviews.length > 1 && (
-                        <button
-                          type="button"
-                          title="Add another multiview"
-                          onClick={() => removeNewMultiview(index)}
-                        >
-                          <IconTrash
-                            className={`ml-4 text-button-delete hover:text-red-400`}
-                          />
-                        </button>
-                      )}
-                      {multiviews.length === index + 1 && (
-                        <button
-                          type="button"
-                          title="Add another multiview"
-                          onClick={() => addNewMultiview(singleItem)}
-                        >
-                          <IconPlus className="mr-2 text-green-400 hover:text-green-200" />
-                        </button>
-                      )}
-                    </div>
-                  </div>
-                </div>
-              );
-            })}
         </div>
-        {!!layoutModalOpen && (
-        <MultiviewLayoutSettings
-          production={production}
-          selectedMultiviewLayout={selectedMultiviewLayout}
-          setNewMultiviewPreset={(newLayout) =>
-            setNewMultiviewLayout(newLayout)
-          }
-        />
-      )}
         <div className="text-button-delete text-center">{currentError}</div>
-        <Decision
-        onClose={() => (layoutModalOpen ? closeLayoutModal() : clearInputs())}
-        onSave={() => (layoutModalOpen ? onUpdateLayoutPreset() : onSave())}
-      />
+        <Decision onClose={() => onClose()} onSave={onSave} />
       </div>
     </Modal>
   );
