+++ src/components/startProduction/ConfigureOutputButton.tsx
@@ -6,20 +6,17 @@ import { Preset } from '../../interfaces/preset';
 import { useTranslate } from '../../i18n/useTranslate';
 import { Button } from '../button/Button';
 import { ConfigureOutputModal } from '../modal/configureOutputModal/ConfigureOutputModal';
-import { Production } from '../../interfaces/production';
 
 type ConfigureOutputButtonProps = {
   preset?: Preset;
   disabled?: boolean;
   updatePreset: (preset: Preset) => void;
-  production: Production | undefined;
 };
 
 export function ConfigureOutputButton({
   preset,
   updatePreset,
-  disabled,
-  production
+  disabled
 }: ConfigureOutputButtonProps) {
   const [modalOpen, setModalOpen] = useState(false);
   const [isDisabled, setIsDisabled] = useState<boolean>(
@@ -61,7 +58,6 @@ export function ConfigureOutputButton({
           preset={preset}
           onClose={toggleConfigModal}
           updatePreset={updatePreset}
-          production={production}
         />
       )}
     </>
